requirements.txt
django>=4.2
google-genai
python-dotenv
gunicorn
# psycopg2-binary  # add this only if you use PostgreSQL
